Add ability to control how MultiMap does joins
In the current implementation, MultiMap will join the left side using
its key field with the right side using the field defined by the
IForeignKeyPropertyResolver resolver.

We have a scenario where a join between left and right is on non-key
fields for both left and right.  The current implementation cannot
cater for this as the left side is always the key field.

To allow a caller to override this we introduce the
IMultiMapJoinResolver interface with associated code.  The caller can
now control what the left and right column names should be.

In our code we introduce a JoinAttribute that you can put on the
left-side property to indicate the property names for both left and
right to use for the joins.  We then implement a
MultiMapJoinAttributeResolver that checks for this attribute and pulls
out the left and right key fields defined, or falls back to the default.