SqlBuilder - Allow caller to use delegate to generate Where
New overloads to Where, OrWhere, AndWhere to use a delegate to generate
the where expression.

```Func<ISqlBuilder, string>```

Example Scenario:

We accept a user input for the Where expression as a string.

e.g.  "/api/customers?filter=Name like '% Smith' and CreatedAt > '2021-01-01' and Status in [0,1]

We parse this input (using Superpower), then convert each field name
to the column name using the ISqlBuilder instance.

e.g. Above gets converted to:

[Customers].[Name] like '% Smith'
and [Customers].[CreatedAt] > '2021-01-01'
and [Customers].[Status] in (0,1)

Where the field names in the filter expression are generated from the
looking up the field names using the Type and then using
Resolvers.Column that requires ISqlBuilder instance.

The resulting expression is returned to SqlExpression.Where